replace init preflight with action onDeploy.before
