+++ src/pkg/utils/network.go
@@ -13,6 +13,7 @@ import (
 	"net/url"
 	"os"
 	"path"
+	"regexp"
 	"strconv"
 	"strings"
 
@@ -107,6 +108,30 @@ func GetAvailablePort() (int, error) {
 		return 0, err
 	}
 	return port, err
+}
+
+// IsValidHostName returns a boolean indicating if the hostname of the host machine is valid according to https://www.ietf.org/rfc/rfc1123.txt.
+func IsValidHostName() bool {
+	// Quick & dirty character validation instead of a complete RFC validation since the OS is already allowing it
+	hostname, err := os.Hostname()
+
+	if err != nil {
+		return false
+	}
+
+	return validHostname(hostname)
+}
+
+func validHostname(hostname string) bool {
+	// Explanation: https://regex101.com/r/zUGqjP/1/
+	rfcDomain := regexp.MustCompile(`^[a-zA-Z0-9\-.]+$`)
+	// Explanation: https://regex101.com/r/vPGnzR/1/
+	localhost := regexp.MustCompile(`\.?localhost$`)
+	isValid := rfcDomain.MatchString(hostname)
+	if isValid {
+		isValid = !localhost.MatchString(hostname)
+	}
+	return isValid
 }
 
 func httpGetFile(url string, destinationFile *os.File) {
